refactor: move GarbageInputValueCheck to own file
